refactor: move components and utils out of app folder
