+++ frontend/package.json
@@ -12,8 +12,10 @@
     "doppler": "doppler secrets download --no-file --format env-no-quotes > .env"
   },
   "dependencies": {
+    "@radix-ui/react-label": "^2.1.2",
     "@radix-ui/react-select": "^2.1.6",
     "@radix-ui/react-slot": "^1.1.2",
+    "@radix-ui/react-switch": "^1.1.3",
     "@radix-ui/react-tabs": "^1.1.3",
     "@stomp/stompjs": "^7.0.0",
     "@tailwindcss/oxide": "^4.0.11",
@@ -28,6 +30,7 @@
     "openapi-fetch": "^0.13.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-textfit": "^1.1.1",
     "tailwind-merge": "^3.0.2",
     "tailwindcss": "^4.0.11",
     "tailwindcss-animate": "^1.0.7",
@@ -40,6 +43,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@types/react-textfit": "^1.1.4",
     "autoprefixer": "^10.4.20",
     "eslint": "^9",
     "eslint-config-next": "15.2.1",
