+++ frontend/src/app/coin/[market]/page.tsx
@@ -1,15 +1,16 @@
 import ClientPage from "@/app/coin/[market]/ClientPage";
-import WebSocketProvider from "@/app/context/WebSocketContext";
+import WebSocketProvider from "@/context/WebSocketContext";
 
 export default function Page() {
   return (
     <>
-      <WebSocketProvider subscriptions={
-        [
+      <WebSocketProvider
+        subscriptions={[
           { type: "ticker", markets: [] },
-          { type: "trade", markets: [] }
-        ]
-      }>
+          { type: "trade", markets: [] },
+          { type: "orderbook", markets: [] },
+        ]}
+      >
         <ClientPage />
       </WebSocketProvider>
     </>
