--- frontend/src/app/components/Pagination.tsx
@@ -1,49 +1,0 @@
-import { useState } from "react";
-
-interface PaginationProps {
-    totalItems: number;
-    itemsPerPage: number;
-    currentPage: number;
-    onPageChange: (page: number) => void;
-}
-
-export default function Pagination({
-    totalItems,
-    itemsPerPage,
-    currentPage,
-    onPageChange,
-}: PaginationProps) {
-    const totalPages = Math.ceil(totalItems / itemsPerPage);
-    const pages = Array.from({ length: totalPages }, (_, i) => i + 1);
-
-    return (
-        <div className="flex justify-center space-x-2 mt-4">
-            <button
-                className="px-3 py-1 border rounded disabled:opacity-50"
-                onClick={() => onPageChange(currentPage - 1)}
-                disabled={currentPage === 1}
-            >
-                이전
-            </button>
-
-            {pages.map((page) => (
-                <button
-                    key={page}
-                    className={`px-3 py-1 border rounded ${currentPage === page ? "bg-blue-500 text-white" : "bg-white"
-                        }`}
-                    onClick={() => onPageChange(page)}
-                >
-                    {page}
-                </button>
-            ))}
-
-            <button
-                className="px-3 py-1 border rounded disabled:opacity-50"
-                onClick={() => onPageChange(currentPage + 1)}
-                disabled={currentPage === totalPages}
-            >
-                다음
-            </button>
-        </div>
-    );
-}
