+++ frontend/src/app/layout.tsx
@@ -2,9 +2,8 @@ import type React from "react";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "@/app/globals.css";
-import Header from "@/app/components/Header";
-import { AuthProvider } from "@/app/context/auth-context";
-
+import { AuthProvider } from "@/context/AuthContext";
+import Header from "@/components/Header";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -14,18 +13,18 @@ export const metadata: Metadata = {
 };
 
 export default function RootLayout({
-                                     children,
-                                   }: {
+  children,
+}: {
   children: React.ReactNode;
 }) {
   return (
-      <html lang="en">
+    <html lang="en">
       <body className={inter.className}>
-      <AuthProvider>
-        <Header />
-        <main className="container mx-auto px-4 py-8">{children}</main>
-      </AuthProvider>
+        <AuthProvider>
+          <Header />
+          <main className="container mx-auto px-4 py-8">{children}</main>
+        </AuthProvider>
       </body>
-      </html>
+    </html>
   );
 }
