+++ frontend/src/app/user/login/page.tsx
@@ -2,84 +2,104 @@
 
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { useAuth } from "../../context/auth-context"; // useAuth 가져오기
+import { useAuth } from "@/context/AuthContext";
 
 export default function LoginPage() {
-    const router = useRouter();
-    const { accessToken, setAccessToken } = useAuth(); // AuthContext에서 토큰 가져오기
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
+  const router = useRouter();
+  const { accessToken, setAccessToken } = useAuth(); // AuthContext에서 토큰 가져오기
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
 
-    // ✅ 로그인 상태라면 자동 리다이렉트
-    useEffect(() => {
-        if (accessToken) {
-            router.push("/");
+  // ✅ 로그인 상태라면 자동 리다이렉트
+  useEffect(() => {
+    if (accessToken) {
+      router.push("/");
+    }
+  }, [accessToken, router]);
+
+  const handleLogin = async (e: React.FormEvent) => {
+    e.preventDefault();
+
+    try {
+      const response = await fetch(
+        process.env.NEXT_PUBLIC_API_URL + "/api/auth/login",
+        {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ email, password }),
+          credentials: "include",
         }
-    }, [accessToken, router]);
+      );
 
-    const handleLogin = async (e: React.FormEvent) => {
-        e.preventDefault();
-
-        try {
-            const response = await fetch(process.env.NEXT_PUBLIC_API_URL +"/api/auth/login", {
-                method: "POST",
-                headers: { "Content-Type": "application/json" },
-                body: JSON.stringify({ email, password }),
-                credentials: "include",
-            });
-
-            if (response.ok) {
-                const authHeader = response.headers.get("Authorization");
-                if (authHeader?.startsWith("Bearer ")) {
-                    const token = authHeader.slice(7);
-                    setAccessToken(token); // 인메모리에 저장
-                    console.log("Access Token 저장됨:", token);
-                }
-
-                alert("로그인 성공!");
-                router.push("/");
-            } else {
-                const errorData = await response.json();
-                alert(`로그인 실패: ${errorData.message}`);
-            }
-        } catch (error) {
-            console.error("Login error:", error);
-            alert("로그인 중 오류가 발생했습니다.");
+      if (response.ok) {
+        const authHeader = response.headers.get("Authorization");
+        if (authHeader?.startsWith("Bearer ")) {
+          const token = authHeader.slice(7);
+          setAccessToken(token); // 인메모리에 저장
+          console.log("Access Token 저장됨:", token);
         }
-    };
 
-    return (
-        <div className="min-h-screen flex flex-col items-center justify-center bg-gray-50">
-            <div className="mb-6 text-2xl font-bold">LOGO</div>
-            <div className="w-full max-w-md bg-white p-8 rounded-md shadow">
-                <h1 className="text-2xl font-bold mb-4">환영합니다</h1>
-                <form className="space-y-4" onSubmit={handleLogin}>
-                    <div>
-                        <label className="block text-sm font-medium text-gray-700 mb-1">이메일 주소</label>
-                        <input type="email" required value={email} onChange={(e) => setEmail(e.target.value)}
-                               className="w-full border border-gray-300 rounded px-3 py-2" />
-                    </div>
-                    <div>
-                        <label className="block text-sm font-medium text-gray-700 mb-1">비밀번호</label>
-                        <input type="password" required value={password} onChange={(e) => setPassword(e.target.value)}
-                               className="w-full border border-gray-300 rounded px-3 py-2" />
-                    </div>
-                    <button type="submit" className="w-full bg-blue-600 text-white py-2 rounded hover:bg-blue-700">
-                        로그인
-                    </button>
-                </form>
+        alert("로그인 성공!");
+        router.push("/");
+      } else {
+        const errorData = await response.json();
+        alert(`로그인 실패: ${errorData.message}`);
+      }
+    } catch (error) {
+      console.error("Login error:", error);
+      alert("로그인 중 오류가 발생했습니다.");
+    }
+  };
 
-                {/* ✅ 회원가입 버튼 추가 */}
-                <div className="mt-4 text-center">
-                    <span className="text-gray-600 text-sm">계정이 없으신가요?</span>
-                    <button
-                        onClick={() => router.push("/user/signup")}
-                        className="ml-2 text-blue-600 font-semibold hover:underline"
-                    >
-                        회원가입
-                    </button>
-                </div>
-            </div>
+  return (
+    <div className="min-h-screen flex flex-col items-center justify-center bg-gray-50">
+      <div className="mb-6 text-2xl font-bold">LOGO</div>
+      <div className="w-full max-w-md bg-white p-8 rounded-md shadow">
+        <h1 className="text-2xl font-bold mb-4">환영합니다</h1>
+        <form className="space-y-4" onSubmit={handleLogin}>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              이메일 주소
+            </label>
+            <input
+              type="email"
+              required
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              className="w-full border border-gray-300 rounded px-3 py-2"
+            />
+          </div>
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              비밀번호
+            </label>
+            <input
+              type="password"
+              required
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              className="w-full border border-gray-300 rounded px-3 py-2"
+            />
+          </div>
+          <button
+            type="submit"
+            className="w-full bg-blue-600 text-white py-2 rounded hover:bg-blue-700"
+          >
+            로그인
+          </button>
+        </form>
+
+        {/* ✅ 회원가입 버튼 추가 */}
+        <div className="mt-4 text-center">
+          <span className="text-gray-600 text-sm">계정이 없으신가요?</span>
+          <button
+            onClick={() => router.push("/user/signup")}
+            className="ml-2 text-blue-600 font-semibold hover:underline"
+          >
+            회원가입
+          </button>
         </div>
-    );
+      </div>
+    </div>
+  );
 }
