+++ frontend/src/lib/utils/index.ts
@@ -1,0 +1,9 @@
+export * from "./mock-data.util";
+export * from "./parse-token.util";
+
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
